Fix bug around section in [rocq-get-deps].
The issue came from the fact that the REPL reports new constants and
inductives using kernel names, which do not include the section name
in the path. To circumvent the problem, we run the queries after the
whole file is processed (and hence, all sections closed).

It is unclear how to better fix the underlying issue. One option may
be to change Rocq's behaviour, and have it not include section names
in paths (making them match kernel names).